% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsr-package.R
\docType{package}
\name{statsr-package}
\alias{statsr}
\alias{statsr-package}
\title{statsr: Statistical Methods Functions in R}
\description{
Collection of R functions for the Statistical Methods course.
}
\author{
\strong{Maintainer}: Kairzhan Berkaliyev \email{sara280273@gmail.com}

Authors:
\itemize{
  \item Kairzhan Berkaliyev \email{sara280273@gmail.com}
}

}
\keyword{internal}
